+++ .nvmrc
@@ -1,1 +1,1 @@
-v16
+v18
