+++ jest.config.js
@@ -189,4 +189,4 @@ const config = {
   // watchman: true,
 };
 
-module.exports = config;
+export default config;
